Proposal for global tags refactoring
